+++ src/Aspire.Dashboard/Model/ManageData/ManageDataGridItem.cs
@@ -1,0 +1,18 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Aspire.Dashboard.Model.ManageData;
+
+/// <summary>
+/// Represents an item in the manage data grid, which can be either a resource row or a nested data row.
+/// </summary>
+public sealed class ManageDataGridItem
+{
+    public ResourceDataRow? ResourceRow { get; init; }
+    public DataRow? NestedRow { get; init; }
+    public ResourceViewModel? ParentResource { get; init; }
+    public int Depth { get; init; }
+
+    public bool IsResourceRow => ResourceRow is not null;
+    public bool IsNestedRow => NestedRow is not null;
+}
